43: Alter logic for start id for new global target
With path being list list of coordinates, the logic was treating the coord
as an integer index. This change adds a vertex lookup on the last point in
the path, which is the last global target (and the PP coordinate of our
current pose).

Replace searching for the vertex ID in the path by performing a lookup on
last path.

Also make improvements to logging to better follow paths.